扣除用户Tokens

+++ src/main/java/org/mizoguchi/misaki/common/constant/SystemConstant.java
@@ -1,6 +1,6 @@
 package org.mizoguchi.misaki.common.constant;
 
 public class SystemConstant {
-    public static final Long CHECK_IN_TOKEN_AMOUNT= 10000L;
+    public static final Integer CHECK_IN_TOKEN_AMOUNT= 10000;
     public static final Integer CHECK_IN_CRYSTAL_AMOUNT= 60;
 }
