+++ src/main/java/org/mizoguchi/misaki/pojo/entity/User.java
@@ -39,7 +39,7 @@ public class User {
 
     private LocalDateTime lastLoginTime;
 
-    private Long token;
+    private Integer token;
 
     private Integer crystal;
